ia_de_música/pyproject.toml
[project]
name = "vamoo-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
